+++ packages/react/src/react/views/index.ts
@@ -1,18 +1,10 @@
+export * from './collections/context'
 export * from './collections/create'
 export * from './collections/create.client'
 export * from './collections/home'
 export * from './collections/layouts/collection-form-layout'
 export * from './collections/layouts/collection-layout'
-export * from './collections/list/banner'
-export * from './collections/list/hooks'
-export * from './collections/list/list.client'
-export * from './collections/list/table'
-export * from './collections/list/table/pagination'
-export * from './collections/list/toolbar'
-export * from './collections/list/toolbar/create'
-export * from './collections/list/toolbar/delete'
-export * from './collections/list/toolbar/filter'
-export * from './collections/list/toolbar/search'
+export * from './collections/list'
 export * from './collections/one'
 export * from './collections/types'
 export * from './collections/update'
